requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.3.0
django-filter==24.2
djangorestframework-simplejwt==5.3.1
celery==5.4.0
